requirements.txt
# Beach Monitor MVP Dependencies
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=8.3.0
PyYAML>=6.0
python-dotenv>=0.19.0

# Agent dependencies
httpx==0.25.2
langchain==0.1.16
langchain-core==0.1.45
langchain-openai==0.1.3
langgraph==0.0.37

# UI dependencies
streamlit>=1.28.0

# Optional: for enhanced logging
loguru>=0.7.0

# YouTube capture
google-api-python-client>=2.100.0
yt-dlp>=2023.10.16

# S3
boto3>=1.28.0
